I found crazy things like that the call to `setvartype!` is ~15% of the total runtime of POWDER (a farming model) and consumes 20% of the memory since it has to do (with Gurobi)  [this dictionary lookup](https://github.com/JuliaOpt/Gurobi.jl/blob/master/src/GurobiSolverInterface.jl#L460) for every variable and it [copies the `colCat` vector](https://github.com/JuliaOpt/JuMP.jl/blob/release-0.18/src/solvers.jl#L390) every solve. Why this is maddening is that we never change the variable type...

This is kind of related to https://github.com/JuliaOpt/JuMP.jl/issues/1110

I'm not sure if this should be merged as it adds some global variables and makes things even more brittle w.r.t. JuMP.

cc @joaquimg for interest

Without these modifications:
```
───────────────────────────────────────────────────────────────────────────────
      Timing statistics               Time                   Allocations
                               ──────────────────────   ───────────────────────
       Tot / % measured:           28252s / 100%            1423GiB / 100%
 Section               ncalls     time   %tot     avg     alloc   %tot      avg
───────────────────────────────────────────────────────────────────────────────
 Solve                      1   28251s   100%  28251s   1423GiB  100%   1423GiB
   Iteration Phase        500   28250s   100%   56.5s   1423GiB  100%   2.85GiB
     Backward Pass        500   28113s   100%   56.2s   1419GiB  100%   2.84GiB
       JuMP.solve       8.64M   27987s  99.1%  3.24ms   1410GiB  99.1%   171KiB
         optimize!      8.64M   23253s  82.3%  2.69ms         -  0.00%        -
         prep JuMP ...  8.64M    4404s  15.6%   510μs   1037GiB  72.8%   126KiB
         getsolution    8.64M     313s  1.11%  36.2μs    372GiB  26.2%  45.2KiB
     Forward Pass         500     137s  0.48%   274ms   4.13GiB  0.29%  8.46MiB
       JuMP.solve       26.0k     134s  0.48%  5.16ms   3.90GiB  0.27%   157KiB
         optimize!      26.0k     118s  0.42%  4.53ms    778KiB  0.00%        -
         prep JuMP ...  26.0k    15.3s  0.05%   590μs   2.86GiB  0.20%   116KiB
         getsolution    26.0k    1.02s  0.00%  39.3μs   1.03GiB  0.07%  41.4KiB 
───────────────────────────────────────────────────────────────────────────────
```